# openapi.yml
openapi: 3.0.0
info:
  title: School Management System API
  version: 1.0.0
  description: API documentation for School Management System
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.production.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          example: false
        message:
          type: string
          example: An error occurred
      required:
        - success
        - message
...
    SectionInput:
      type: object
      properties:
        name:
          type: string
          example: Section A
        classId:
          type: string
          example: '101'
      required:
        - name
        - classId
    Section:
      type: object
      properties:
        sectionId:
          type: string
          example: '1'
        name:
          type: string
          example: Section A
        classId:
          type: string
          example: '101'
      required:
        - name
        - classId
    SectionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Section'
        message:
          type: string
          example: Section created successfully
    SubjectInput:
      type: object
      properties:
        subjectName:
          type: string
          example: Mathematics
        subjectCode:
          type: string
          example: MATH101
      required:
        - subjectName
        - subjectCode
    Subject:
      type: object
      properties:
        subjectId:
          type: integer
          example: 1
        subjectName:
          type: string
          example: Mathematics
        subjectCode:
          type: string
          example: MATH101
      required:
        - subjectName
        - subjectCode
    SubjectResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Subject'
        message:
          type: string
          example: Subject created successfully
...
security:
  - bearerAuth: []
paths:
...
  /sections:
    post:
      summary: Create a new section
      tags:
        - Sections
      security:
        - bearerAuth: []
      requestBody:
        description: Payload to create a new section.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionInput'
            examples:
              sectionInput:
                summary: New section input example
                value:
                  name: Section A
                  classId: '101'
      responses:
        '201':
          description: Section created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sections/class/{classId}:
    get:
      summary: Get all sections for a class
      tags:
        - Sections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          description: Unique identifier for the class.
          schema:
            type: string
          example: '101'
      responses:
        '200':
          description: List of sections
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Section'
                  message:
                    type: string
                    example: Sections retrieved successfully
        '404':
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sections/{sectionId}:
    get:
      summary: Get section by ID
      tags:
        - Sections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sectionId
          required: true
          description: Unique identifier for the section.
          schema:
            type: string
          example: '1'
      responses:
        '200':
          description: Section details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a section
      tags:
        - Sections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sectionId
          required: true
          description: Unique identifier for the section to update.
          schema:
            type: string
          example: '1'
      requestBody:
        description: Updated section data.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionInput'
            examples:
              sectionInput:
                summary: Update section input example
                value:
                  name: Section A Updated
                  classId: '102'
      responses:
        '200':
          description: Section updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a section
      tags:
        - Sections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sectionId
          required: true
          description: Unique identifier for the section to delete.
          schema:
            type: string
          example: '1'
      responses:
        '200':
          description: Section deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Section deleted successfully
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subjects:
    post:
      summary: Create a new subject
      tags:
        - Subjects
      security:
        - bearerAuth: []
      requestBody:
        description: Subject information for creating a new subject
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectInput'
            examples:
              newSubject:
                summary: New subject input example
                value:
                  subjectName: Physics
                  subjectCode: PHYS101
      responses:
        '201':
          description: Subject created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          description: Invalid input or error adding subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all subjects
      tags:
        - Subjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subjects retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '400':
          description: Error retrieving subjects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subjects/{subjectId}:
    get:
      summary: Get a subject by ID
      tags:
        - Subjects
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subjectId
          required: true
          description: Numeric ID of the subject to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Subject retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          description: Invalid subject ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subject not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a subject by ID
      tags:
        - Subjects
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subjectId
          description: Numeric ID of the subject to update
          required: true
          schema:
            type: integer
      requestBody:
        description: Updated subject data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectInput'
            examples:
              updateSubject:
                summary: Updated subject input example
                value:
                  subjectName: Advanced Physics
                  subjectCode: PHYS201
      responses:
        '200':
          description: Subject updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          description: Invalid subject ID or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subject not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a subject by ID
      tags:
        - Subjects
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subjectId
          description: Numeric ID of the subject to delete
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Subject deleted successfully. No content returned.
        '400':
          description: Invalid subject ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subject not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...